11 安装第三方包（演示安装 Intl 包）

+++ lib/demo/SliderDemo.dart
@@ -28,7 +28,7 @@ class _SliderDemoState extends State<SliderDemo> {
         child: Column(
           mainAxisAlignment: MainAxisAlignment.center,
           children: <Widget>[
-            Text(_sliderItem.toInt().toString()),
+            Text(_sliderItem.toString()),
             Slider(
               value: _sliderItem,
               onChanged: _handleOnChanged,
